[CP-2130] Never ending files upload
Corrected check to see if a device has enough free space to upload a file.

+++ module-services/service-desktop/endpoints/include/endpoints/filesystem/FS_Helper.hpp
@@ -42,7 +42,7 @@ namespace sdesktop::endpoints
         auto startGetFile(Context &context) const -> ResponseContext;
         auto getFileChunk(Context &context) const -> ResponseContext;
 
-        auto getFreeSpaceForUserFilesMiB() const -> unsigned long;
+        auto getFreeSpaceForUserFilesMiB() const -> float;
 
         auto startSendFile(Context &context) const -> ResponseContext;
         auto sendFileChunk(Context &context) const -> ResponseContext;
